# internal/server/app_plan.go
package server

import (
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"os"
	"path/filepath"
	"strings"
	"time"

	"github.com/google/uuid"

	"tofi-core/internal/provider"
	"tofi-core/internal/workspace"
)

// AppPlan represents a temporary app plan generated by AI.
type AppPlan struct {
	PlanID    string             `json:"plan_id"`
	Prompt    string             `json:"prompt"`
	App       json.RawMessage    `json:"app"`
	Messages  []provider.Message `json:"messages"`
	CreatedAt time.Time          `json:"created_at"`
	ExpiresAt time.Time          `json:"expires_at"`
}

const planTTL = 1 * time.Hour

// generatePlanID creates a short plan ID.
func generatePlanID() string {
	return "plan-" + uuid.New().String()[:8]
}

// planFilePath returns the file path for a plan.
func (s *Server) planFilePath(userID, planID string) string {
	return filepath.Join(s.workspace.PlansDir(userID), planID+".json")
}

// readPlan reads a plan from disk.
func (s *Server) readPlan(userID, planID string) (*AppPlan, error) {
	data, err := os.ReadFile(s.planFilePath(userID, planID))
	if err != nil {
		return nil, err
	}
	var plan AppPlan
	if err := json.Unmarshal(data, &plan); err != nil {
		return nil, err
	}
	return &plan, nil
}

// writePlan writes a plan to disk.
func (s *Server) writePlan(userID string, plan *AppPlan) error {
	data, err := json.MarshalIndent(plan, "", "  ")
	if err != nil {
		return err
	}
	return os.WriteFile(s.planFilePath(userID, plan.PlanID), data, 0644)
}

// deletePlan removes a plan file.
func (s *Server) deletePlan(userID, planID string) error {
	return os.Remove(s.planFilePath(userID, planID))
}

// buildPlanSystemPrompt creates the system prompt with user timezone injected.
func buildPlanSystemPrompt(timezone string) string {
	tzClause := ""
	if timezone != "" {
		tzClause = fmt.Sprintf(`

User's timezone: %s
When generating schedule_rules, use this timezone unless the user explicitly specifies a different one.`, timezone)
	}

	return fmt.Sprintf(`You are an App Planner for Tofi — a platform where users create AI Apps that run on schedules or on-demand.

Your job: convert natural language descriptions into a complete app configuration.

Output ONLY valid JSON in this exact format (no markdown, no explanation):
{
  "id": "kebab-case-id",
  "name": "Display Name",
  "description": "One-line description of what this app does",
  "prompt": "Detailed task instructions for the AI agent that will execute this app. Write as if briefing a capable employee who knows nothing about the user's context. Be specific about what to do, what format to use, and what to deliver.",
  "model": "",
  "skills": [],
  "schedule_rules": {
    "entries": [
      {
        "time": "09:00",
        "repeat": {"type": "daily"},
        "enabled": true
      }
    ],
    "timezone": "America/New_York"
  }
}

Rules:
- id: kebab-case, short and meaningful (e.g. "daily-email-digest", "stock-monitor")
- prompt: detailed, professional task instructions — this is the ONLY thing the executing agent sees
- schedule_rules: include ONLY if the user mentions a time or schedule. Omit entirely if no schedule is implied.
- skills: recommend relevant skills if the task needs web search, API calls, etc. Use [] if none needed.
- model: leave empty string "" (system will auto-select)
- If this is a follow-up iteration, modify the previous plan based on the new request while preserving unchanged parts.%s`, tzClause)
}

// handlePlanApp POST /api/v1/plans
// New plan: POST /api/v1/plans  {"prompt": "..."}
// Iterate:  POST /api/v1/plans?id=plan-xxx  {"prompt": "..."}
func (s *Server) handlePlanApp(w http.ResponseWriter, r *http.Request) {
	userID := r.Context().Value(UserContextKey).(string)

	var req struct {
		Prompt string `json:"prompt"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeJSONError(w, http.StatusBadRequest, ErrBadRequest, "invalid request body", "")
		return
	}
	if req.Prompt == "" {
		writeJSONError(w, http.StatusBadRequest, ErrBadRequest, "prompt is required", "")
		return
	}

	// Resolve model + key
	model, apiKey, _, err := s.resolveModelAndKey(userID, "")
	if err != nil {
		writeJSONError(w, http.StatusBadRequest, ErrBadRequest,
			fmt.Sprintf("no API key available: %v", err),
			"Set an AI key first: PUT /api/v1/user/settings/ai-key")
		return
	}

	// Read timezone from header or query param
	timezone := r.Header.Get("X-Timezone")
	if timezone == "" {
		timezone = r.URL.Query().Get("timezone")
	}

	// Build messages for LLM call
	var messages []provider.Message
	var planID string

	existingPlanID := r.URL.Query().Get("id")
	if existingPlanID != "" {
		// Iterate mode: load existing plan + conversation history
		existing, err := s.readPlan(userID, existingPlanID)
		if err != nil {
			writeJSONError(w, http.StatusNotFound, ErrNotFound, "plan not found or expired", "")
			return
		}
		if time.Now().After(existing.ExpiresAt) {
			s.deletePlan(userID, existingPlanID)
			writeJSONError(w, http.StatusNotFound, ErrNotFound, "plan expired", "")
			return
		}
		messages = existing.Messages
		planID = existing.PlanID
	} else {
		// New plan
		planID = generatePlanID()
	}

	// Append new user message
	messages = append(messages, provider.Message{
		Role:    "user",
		Content: req.Prompt,
	})

	// Call LLM (non-streaming)
	p, err := provider.NewForModel(model, apiKey)
	if err != nil {
		writeJSONError(w, http.StatusInternalServerError, ErrInternal,
			fmt.Sprintf("failed to create provider: %v", err), "")
		return
	}

	llmResp, err := p.Chat(r.Context(), &provider.ChatRequest{
		Model:    model,
		System:   buildPlanSystemPrompt(timezone),
		Messages: messages,
	})
	if err != nil {
		writeJSONError(w, http.StatusInternalServerError, ErrInternal,
			fmt.Sprintf("LLM call failed: %v", err), "")
		return
	}

	// Parse LLM output as JSON
	cleaned := cleanJSONResponse(llmResp.Content)
	var appSpec json.RawMessage
	if err := json.Unmarshal([]byte(cleaned), &appSpec); err != nil {
		writeJSONError(w, http.StatusInternalServerError, ErrInternal,
			"AI returned invalid JSON, please try again",
			fmt.Sprintf("Raw response: %s", llmResp.Content))
		return
	}

	// Append assistant response to history
	messages = append(messages, provider.Message{
		Role:    "assistant",
		Content: cleaned,
	})

	// Build plan
	now := time.Now()
	plan := &AppPlan{
		PlanID:    planID,
		Prompt:    req.Prompt,
		App:       appSpec,
		Messages:  messages,
		CreatedAt: now,
		ExpiresAt: now.Add(planTTL),
	}

	// Write to disk
	if err := s.writePlan(userID, plan); err != nil {
		writeJSONError(w, http.StatusInternalServerError, ErrInternal,
			fmt.Sprintf("failed to save plan: %v", err), "")
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(plan)
}

// handleGetPlan GET /api/v1/plans/{planId}
func (s *Server) handleGetPlan(w http.ResponseWriter, r *http.Request) {
	userID := r.Context().Value(UserContextKey).(string)
	planID := r.PathValue("planId")

	plan, err := s.readPlan(userID, planID)
	if err != nil {
		writeJSONError(w, http.StatusNotFound, ErrNotFound, "plan not found", "")
		return
	}
	if time.Now().After(plan.ExpiresAt) {
		s.deletePlan(userID, planID)
		writeJSONError(w, http.StatusNotFound, ErrNotFound, "plan expired", "")
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(plan)
}

// handleApprovePlan POST /api/v1/plans/{planId}/approve
// Creates the app from the plan, with optional field overrides.
func (s *Server) handleApprovePlan(w http.ResponseWriter, r *http.Request) {
	userID := r.Context().Value(UserContextKey).(string)
	planID := r.PathValue("planId")

	plan, err := s.readPlan(userID, planID)
	if err != nil {
		writeJSONError(w, http.StatusNotFound, ErrNotFound, "plan not found", "")
		return
	}
	if time.Now().After(plan.ExpiresAt) {
		s.deletePlan(userID, planID)
		writeJSONError(w, http.StatusNotFound, ErrNotFound, "plan expired", "")
		return
	}

	// Parse the plan's app spec into a mutable map
	var appSpec map[string]json.RawMessage
	if err := json.Unmarshal(plan.App, &appSpec); err != nil {
		writeJSONError(w, http.StatusInternalServerError, ErrInternal, "corrupted plan data", "")
		return
	}

	// Parse optional overrides from request body
	if r.Body != nil && r.ContentLength > 0 {
		var overrides map[string]json.RawMessage
		if err := json.NewDecoder(r.Body).Decode(&overrides); err == nil {
			for k, v := range overrides {
				appSpec[k] = v
			}
		}
	}

	// Extract fields from merged spec
	getString := func(key string) string {
		raw, ok := appSpec[key]
		if !ok {
			return ""
		}
		var s string
		json.Unmarshal(raw, &s)
		return s
	}
	getStringSlice := func(key string) []string {
		raw, ok := appSpec[key]
		if !ok {
			return nil
		}
		var ss []string
		json.Unmarshal(raw, &ss)
		return ss
	}
	getRaw := func(key string) *json.RawMessage {
		raw, ok := appSpec[key]
		if !ok {
			return nil
		}
		return &raw
	}

	appID := getString("id")
	if appID == "" {
		writeJSONError(w, http.StatusBadRequest, ErrBadRequest, "plan is missing app id", "")
		return
	}
	if !isValidAppID(appID) {
		writeJSONError(w, http.StatusBadRequest, ErrBadRequest,
			"app id must be kebab-case", "")
		return
	}

	// Build AgentDef from plan
	def := requestToAgentDef(
		appID,
		getString("name"),
		getString("description"),
		getString("prompt"),
		getString("system_prompt"),
		getString("model"),
		getStringSlice("skills"),
		getRaw("schedule_rules"),
		getRaw("capabilities"),
		nil, // buffer_size
		nil, // renewal_threshold
		getRaw("parameter_defs"),
	)

	// Write agent files
	if s.workspace != nil {
		if err := s.workspace.WriteAgent(userID, def); err != nil {
			writeJSONError(w, http.StatusInternalServerError, ErrInternal,
				fmt.Sprintf("failed to write agent: %v", err), "")
			return
		}
	}

	// Sync to DB
	var record interface{}
	if s.workspaceSync != nil {
		synced, err := s.workspaceSync.SyncAgentToDB(userID, appID)
		if err != nil {
			writeJSONError(w, http.StatusInternalServerError, ErrInternal,
				fmt.Sprintf("failed to sync agent: %v", err), "")
			return
		}
		record = synced
	} else {
		rec := workspace.AgentDefToRecord(userID, def)
		if err := s.db.CreateApp(rec); err != nil {
			writeJSONError(w, http.StatusInternalServerError, ErrInternal,
				fmt.Sprintf("failed to create app: %v", err), "")
			return
		}
		record = rec
	}

	// Clean up plan file
	s.deletePlan(userID, planID)

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(record)
}

// handleDeletePlan DELETE /api/v1/plans/{planId}
func (s *Server) handleDeletePlan(w http.ResponseWriter, r *http.Request) {
	userID := r.Context().Value(UserContextKey).(string)
	planID := r.PathValue("planId")

	if err := s.deletePlan(userID, planID); err != nil {
		writeJSONError(w, http.StatusNotFound, ErrNotFound, "plan not found", "")
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"status": "deleted",
	})
}

// startPlanCleanup launches a goroutine that periodically removes expired plan files.
func (s *Server) startPlanCleanup() {
	go func() {
		ticker := time.NewTicker(10 * time.Minute)
		defer ticker.Stop()
		for range ticker.C {
			s.cleanExpiredPlans()
		}
	}()
}

func (s *Server) cleanExpiredPlans() {
	usersDir := filepath.Join(s.config.HomeDir, "users")
	userEntries, err := os.ReadDir(usersDir)
	if err != nil {
		return
	}

	now := time.Now()
	for _, ue := range userEntries {
		if !ue.IsDir() {
			continue
		}
		plansDir := filepath.Join(usersDir, ue.Name(), "plans")
		entries, err := os.ReadDir(plansDir)
		if err != nil {
			continue
		}
		for _, e := range entries {
			if e.IsDir() || !strings.HasSuffix(e.Name(), ".json") {
				continue
			}
			path := filepath.Join(plansDir, e.Name())
			data, err := os.ReadFile(path)
			if err != nil {
				continue
			}
			var plan AppPlan
			if err := json.Unmarshal(data, &plan); err != nil {
				continue
			}
			if now.After(plan.ExpiresAt) {
				os.Remove(path)
				log.Printf("[plan-cleanup] removed expired plan: %s/%s", ue.Name(), e.Name())
			}
		}
	}
}