# internal/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tofi API
  version: 0.1.7
  description: |
    Tofi is an AI App Engine. Create AI-powered apps that run on a schedule
    or on-demand, with results delivered to your preferred channels.

    ## Authentication
    All endpoints (except login) require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Quick Start

    **Step 1: Login**
    ```bash
    curl -X POST https://tofi.sentiosurge.com/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"username": "your-username", "password": "your-password"}'
    ```
    Save the `token` from the response.

    **Step 2: Check your status**
    ```bash
    curl https://tofi.sentiosurge.com/api/v1/auth/me \
      -H "Authorization: Bearer YOUR_TOKEN"
    ```
    This shows your configured AI providers. If `ai_keys` is empty, go to Step 3.

    **Step 3: Set your AI key**
    ```bash
    curl -X PUT https://tofi.sentiosurge.com/api/v1/user/settings/ai-key \
      -H "Authorization: Bearer YOUR_TOKEN" \
      -H "Content-Type: application/json" \
      -d '{"provider": "openai", "api_key": "sk-..."}'
    ```
    Supported providers: `openai`, `anthropic`, `gemini`, `deepseek`, `groq`, `openrouter`

    **Step 4: Start chatting**
    ```bash
    # Create a session
    curl -X POST https://tofi.sentiosurge.com/api/v1/chat/sessions \
      -H "Authorization: Bearer YOUR_TOKEN" \
      -H "Content-Type: application/json" \
...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/plans/{planId}:
    get:
      summary: Get Plan
      description: View a plan's app spec and conversation history.
      operationId: getPlan
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: Plan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '404':
          description: Plan not found or expired
    delete:
      summary: Delete Plan
      description: Discard a plan. Plans also auto-expire after 1 hour.
      operationId: deletePlan
      tags: [Plans]
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: Plan deleted
        '404':
          description: Plan not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via POST /api/v1/auth/login
  parameters:
    PlanId:
      name: planId
      in: path
      required: true
      description: Plan ID (e.g. plan-a1b2c3d4)
      schema:
        type: string
...
  schemas:
    Plan:
      type: object
      properties:
        plan_id:
          type: string
          example: "plan-a1b2c3d4"
        prompt:
          type: string
          example: "Every morning at 9am, summarize tech news"
        app:
          type: object
          description: Complete app spec (same format as POST /apps request body)
        messages:
          type: array
          description: Conversation history for iterating the plan
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
...